First working implementation
Now have the ability to save and then reload games. Format for saves is
textfile with paths and choices separated in CSV format. Implementation
doesn't work propoerly for more complex games and the entire code needs
to be refactored to include a dedicated Tree data structure. But next
will write unit tests for behavior, which should allow next version to
be made with efficiency.